+++ tests/kernel/threads/thread_apis/prj.conf
@@ -1,4 +1,5 @@
 CONFIG_ZTEST=y
 CONFIG_THREAD_MONITOR=y
-CONFIG_HEAP_MEM_POOL_SIZE=256
 CONFIG_THREAD_CUSTOM_DATA=y
+CONFIG_THREAD_STACK_INFO=y
+CONFIG_HEAP_MEM_POOL_SIZE=256
